Add `coverage` command to estimate training data coverage
This commit introduces a new `coverage` command that calculates the coverage of training data by analyzing events within a specified date range and frequency. Adjustments were also made to the `subrange` method in the journal to support datetime inputs and link the new command to the main CLI entry point.